check if file exists before saving

+++ apps/container/src/utils/fs.ts
@@ -45,9 +45,20 @@ export async function writePathContent(dist: string, body: string) {
   try {
     const newPath = path.join(prefix, dist)
     if (!newPath.startsWith(prefix)) return r('ERROR')
+    const isFile = await isValidFile(newPath)
+    if (!isFile) return r('INVALID_PATH')
     await fs.writeFile(newPath, body)
     return r('OK')
   } catch (error) {
     return r('ERROR')
+  }
+}
+
+async function isValidFile(path: string) {
+  try {
+    const stat = await fs.stat(path)
+    return !stat.isDirectory()
+  } catch (error) {
+    return false
   }
 }
